Refactor and restructure rules API
